PUD-668: Setup a clamav deployment and service. (#202)
This is the common method HMPPS teams use for scanning uploaded assets
(before storage in S3). The service will listen for TCP requests at
`manage-recalls-api-clamav.svc` on port 3310.

There is also a cronjob implemented as part of the helm chart that
updates the image used by the clamav pod on a twice daily basis - this
is how the virus definition files are kept up to date.

We will need to use a library to implement the
AV scans on files - [this](https://github.com/cdarras/clamav-client)
seems pretty promising.