PUD-501: Upgrade gotenberg (#483)
This updates gotenberg to version 7.5.0 and changes all the calling code
to the new URLs.

Note the default page sizes and margins in gotenberg 7+ have changed, so
a lot of the changes are to work around that. There is also some janky
spacing in text when read in for the tests (comments added in the code
to the effect). The resulting PDFs look the same to the eye though.

Co-authored-by: Phil Segal <philsegal@users.noreply.github.com>